# backend_go/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkyViewLab Go Control Plane API",
    "version": "1.0.0-gate-b",
    "description": "Machine-checkable contract for the current Go /api/v1 control plane. This contract describes the implemented Gate B boundary; it does not claim that every workbench domain schema or production NO-GO item is complete. All JSON success responses use {data, requestId}; all JSON failures use {error:{code,message}, message, requestId}."
  },
  "servers": [
    {"url": "https://api.example.com", "description": "Production placeholder; deployment supplies the real origin"},
    {"url": "http://127.0.0.1:8080", "description": "Local development"}
  ],
...
  "paths": {
...
    "/api/v1/projects/{id}/jobs": {
      "parameters": [{"$ref": "#/components/parameters/ResourceID"}],
      "get": {"operationId": "listProjectJobs", "tags": ["jobs"], "summary": "List project jobs", "description": "Returns browser-safe job projections without tenant, actor, request input, idempotency, worker identity, or lease data.", "security": [{"sessionCookie": []}], "responses": {"200": {"$ref": "#/components/responses/PublicJobList"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}},
      "post": {"operationId": "createProjectJob", "tags": ["jobs"], "summary": "Create or idempotently replay a server job", "description": "The Idempotency-Key header and body idempotencyKey are alternatives. If both are sent, they must match. Reusing a key with different semantics returns 409. The response is a browser-safe projection and never echoes either the input or idempotency key.", "security": [{"sessionCookie": []}], "parameters": [{"$ref": "#/components/parameters/Origin"}, {"$ref": "#/components/parameters/IdempotencyKey"}], "x-origin-required-for-cookie": true, "x-max-body-bytes": 4194304, "x-max-job-input-bytes": 1048576, "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobCreateRequest"}}}}, "responses": {"200": {"$ref": "#/components/responses/PublicJobRun"}, "201": {"$ref": "#/components/responses/PublicJobRun"}, "400": {"$ref": "#/components/responses/Error400"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "409": {"$ref": "#/components/responses/Error409"}, "413": {"$ref": "#/components/responses/Error413"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
    "/api/v1/jobs/{id}": {
      "parameters": [{"$ref": "#/components/parameters/ResourceID"}],
      "get": {"operationId": "getJob", "tags": ["jobs"], "summary": "Get authorized job", "description": "Returns the browser-safe PublicJob projection.", "security": [{"sessionCookie": []}], "responses": {"200": {"$ref": "#/components/responses/PublicJob"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {"type": "apiKey", "in": "cookie", "name": "skyviewlab_session", "description": "Opaque HttpOnly BFF session. Browser unsafe operations also require an allowlisted Origin."},
      "oidcStateCookie": {"type": "apiKey", "in": "cookie", "name": "skyviewlab_oidc_state", "description": "Short-lived, one-time OIDC correlation cookie; not an authenticated application session."},
      "workerBearer": {"type": "http", "scheme": "bearer", "bearerFormat": "opaque", "description": "A distinct opaque token bound server-side to one worker ID and 1..64 exact tenantId/workspaceId/single-slug/action-allowlist scopes. Empty scopes, wildcards and duplicate boundaries are rejected. Production loads the bounded credential set from WORKER_CREDENTIALS_FILE; short-lived workload JWT or mTLS remains a deployment hardening boundary."}
    },
    "parameters": {
      "Origin": {"name": "Origin", "in": "header", "required": true, "description": "Exactly one browser Origin, matching the configured allowlist byte-for-byte.", "schema": {"type": "string", "format": "uri", "maxLength": 2048}},
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": false, "description": "Exactly one value when present: 1..128 ASCII letters, digits, or -._~:/+=. It must match body idempotencyKey when both are supplied.", "x-single-value": true, "schema": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[A-Za-z0-9._~:/+=-]{1,128}$"}},
...
      "ResourceID": {"name": "id", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 128}},
...
    },
    "schemas": {
      "ObjectMap": {"type": "object", "additionalProperties": true},
      "RequestID": {"type": "string", "minLength": 1, "maxLength": 128},
      "ErrorDetail": {"type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": {"code": {"type": "string", "minLength": 1}, "message": {"type": "string"}}},
      "ErrorEnvelope": {"type": "object", "additionalProperties": false, "required": ["error", "message", "requestId"], "properties": {"error": {"$ref": "#/components/schemas/ErrorDetail"}, "message": {"type": "string"}, "requestId": {"$ref": "#/components/schemas/RequestID"}}},
      "DataEnvelope": {"type": "object", "required": ["data", "requestId"], "properties": {"data": {}, "requestId": {"$ref": "#/components/schemas/RequestID"}}},
...
      "JobStatus": {"type": "string", "enum": ["queued", "running", "succeeded", "failed", "canceled"]},
...
      "PublicJob": {"type": "object", "additionalProperties": false, "description": "Browser-safe job projection. Tenant scope, actor identity, request input, idempotency data, worker identity and lease capabilities are intentionally absent.", "required": ["id", "projectId", "slug", "action", "status", "result", "attempt", "durationMs", "createdAt", "updatedAt"], "properties": {"id": {"type": "string"}, "projectId": {"type": "string"}, "slug": {"type": "string"}, "action": {"type": "string", "maxLength": 80}, "status": {"$ref": "#/components/schemas/JobStatus"}, "result": {"$ref": "#/components/schemas/ObjectMap"}, "error": {"type": "string"}, "attempt": {"type": "integer", "minimum": 0}, "durationMs": {"type": "number", "minimum": 0}, "createdAt": {"type": "string", "format": "date-time"}, "startedAt": {"type": "string", "format": "date-time"}, "updatedAt": {"type": "string", "format": "date-time"}, "finishedAt": {"type": "string", "format": "date-time"}}},
      "PublicRun": {"type": "object", "additionalProperties": false, "description": "Browser-safe server-owned run evidence. Request input is intentionally absent.", "required": ["id", "jobId", "projectId", "slug", "action", "status", "result", "durationMs", "createdAt", "updatedAt"], "properties": {"id": {"type": "string"}, "jobId": {"type": "string"}, "projectId": {"type": "string"}, "slug": {"type": "string"}, "action": {"type": "string"}, "status": {"$ref": "#/components/schemas/JobStatus"}, "result": {"$ref": "#/components/schemas/ObjectMap"}, "error": {"type": "string"}, "durationMs": {"type": "number", "minimum": 0}, "createdAt": {"type": "string", "format": "date-time"}, "updatedAt": {"type": "string", "format": "date-time"}}},
      "PublicJobRun": {"type": "object", "additionalProperties": false, "required": ["job", "run"], "properties": {"job": {"$ref": "#/components/schemas/PublicJob"}, "run": {"$ref": "#/components/schemas/PublicRun"}}},
...
      "JobCreateRequest": {"type": "object", "additionalProperties": false, "required": ["slug", "action"], "properties": {"slug": {"type": "string", "minLength": 1}, "action": {"type": "string", "minLength": 1, "maxLength": 80}, "input": {"$ref": "#/components/schemas/ObjectMap", "default": {}}, "idempotencyKey": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[A-Za-z0-9._~:/+=-]{1,128}$"}}},
...
      "PublicJobEnvelope": {"allOf": [{"$ref": "#/components/schemas/DataEnvelope"}, {"type": "object", "properties": {"data": {"$ref": "#/components/schemas/PublicJob"}}}]},
      "PublicJobListEnvelope": {"allOf": [{"$ref": "#/components/schemas/DataEnvelope"}, {"type": "object", "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/PublicJob"}}}}]},
      "PublicJobRunEnvelope": {"allOf": [{"$ref": "#/components/schemas/DataEnvelope"}, {"type": "object", "properties": {"data": {"$ref": "#/components/schemas/PublicJobRun"}}}]},
...
    },
    "responses": {
...
      "PublicJob": {"description": "Browser-safe job", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicJobEnvelope"}}}},
      "PublicJobList": {"description": "Browser-safe jobs", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicJobListEnvelope"}}}},
      "PublicJobRun": {"description": "Browser-safe job and matching server-owned run", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicJobRunEnvelope"}}}},
...
      "Error400": {"description": "Invalid request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error401": {"description": "Authentication or correlation failed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error403": {"description": "Origin, role, or resource authorization denied", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error404": {"description": "Resource not found within the authorized scope", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error409": {"description": "State, revision, idempotency, or immutable-workflow conflict", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error413": {"description": "Request or encoded job input exceeds the documented limit", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
...
      "Error500": {"description": "Internal persistence, encoding, or audit failure", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
...
      "Error503": {"description": "Required identity, authorization, Worker, database, or compute dependency is unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
...
    }
  }
}